requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.0.3
Flask-SocketIO==5.3.7
frozendict==2.4.4
frozenlist==1.4.1
googlesearch-python==1.2.5
GPUtil==1.4.0
greenlet==3.0.3
groq==0.11.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.16
langchain-core==0.2.38
langchain-text-splitters==0.2.4
langsmith==0.1.114
lxml==5.3.0
MarkupSafe==2.1.5
multidict==6.0.5
multitasking==0.0.11
nltk==3.9.1
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.2
peewee==3.17.6
platformdirs==4.2.2
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
pytz==2024.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
webencodings==0.5.1
Werkzeug==3.0.4
wsproto==1.2.0
yarl==1.9.11
yfinance==0.2.43
youtube-search-python==1.6.6
